+++ src/cliConfig.ts
@@ -14,7 +14,16 @@ export function getGlobalFlags(
 	// Last takes precedence/overrides previous ones
 	return [
 		...(configs.get<string[]>("coder.globalFlags") || []),
-		...["--global-config", escapeCommandArg(configDir)],
+		"--global-config",
+		escapeCommandArg(configDir),
 		...getHeaderArgs(configs),
 	];
+}
+
+/**
+ * Returns SSH flags for the `coder ssh` command from user configuration.
+ */
+export function getSshFlags(configs: WorkspaceConfiguration): string[] {
+	// Make sure to match this default with the one in the package.json
+	return configs.get<string[]>("coder.sshFlags", ["--disable-autostart"]);
 }